[CEF] Merged revisions 6853-6900 from trunk

+++ branches/awesomium/MTA10/loader/CExePatchedStatus.cpp
@@ -131,7 +131,7 @@ SString GetPatchExeAdminReason( bool bUseExeCopy, const SExePatchedStatus& reqSt
 //
 //
 //////////////////////////////////////////////////////////
-uint GetExeFileSize( bool bUseExeCopy )
+uint64 GetExeFileSize( bool bUseExeCopy )
 {
     SString strGTAEXEPath = GetExePathFilename( bUseExeCopy );
     return FileSize( strGTAEXEPath );
